+++ CotEditor/Sources/Settings Window/Panes/FormatSettingsView.swift
@@ -323,8 +323,8 @@ private struct SyntaxListView: View {
                         }
                         
                         let name = url.deletingPathExtension().lastPathComponent
-                        let type = UTType(filenameExtension: url.pathExtension)
                         do {
+                            let type = try url.resourceValues(forKeys: [.contentTypeKey]).contentType
                             try self.manager.importSetting(.url(url), name: name, type: type, overwrite: false)
                         } catch let error as ImportDuplicationError {
                             self.importingError = error