+++ CotEditor/Sources/Text Finder/Multiple Replace/Views/MultipleReplaceListView.swift
@@ -112,8 +112,8 @@ struct MultipleReplaceListView: View {
                         }
                         
                         let name = url.deletingPathExtension().lastPathComponent
-                        let type = UTType(filenameExtension: url.pathExtension)
                         do {
+                            let type = try url.resourceValues(forKeys: [.contentTypeKey]).contentType
                             try self.manager.importSetting(.url(url), name: name, type: type, overwrite: false)
                         } catch let error as ImportDuplicationError {
                             self.importingError = error